requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
pyperclip==1.8.2
selenium==4.15.2
webdriver-manager==4.0.1
Pillow==10.1.0
PyQt5==5.15.10
pywin32==306
lxml_html_clean==0.4.2
